Combine importKey operations of ECDSA and ECDH
